% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_index.R
\name{tidy_index}
\alias{tidy_index}
\alias{tidy_indice}
\title{Tidy a Gapminder indicator data sheet.}
\usage{
tidy_index(filepath)

tidy_indice(filepath)
}
\arguments{
\item{filepath}{Path to a Gapminder indicator data sheet (.csv, .xlsx, or .xls).}
}
\value{
The indicator data sheet as a tidy data frame with columns
   \code{country}, \code{year}, and the indicator name.
}
\description{
Tidy a Gapminder indicator data sheet.
}
\examples{
csv_filepath <- system.file("extdata/life_expectancy_years.csv",
                             package = "tidygapminder")
xlsx_filepath <- system.file(
                 "extdata/agriculture_land.xlsx",
                 package = "tidygapminder")

tidy_index(csv_filepath)
tidy_index(xlsx_filepath)

}
