% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_indice.R
\name{tidy_indice}
\alias{tidy_indice}
\title{Tidy Gapminder indice single data sheet.}
\usage{
tidy_indice(filepath)
}
\arguments{
\item{path}{Absolute path to folder to indice data sheet.}
}
\value{
The indice data sheet as a tidy data frame.
}
\description{
Tidy Gapminder indice single data sheet.
}
\examples{
library(tidygapminder)

tidy_indice(filepath = filename)


}
