% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_indice.R
\name{tidy_indice}
\alias{tidy_indice}
\title{Tidy Gapminder indicator single data sheet.}
\usage{
tidy_indice(filepath)
}
\arguments{
\item{filepath}{Path to indice data sheet.}
}
\value{
The indicator data sheet as a tidy data frame.
}
\description{
Tidy Gapminder indicator single data sheet.
}
\examples{

csv_filepath <- system.file("extdata/life_expectancy_years.csv",
                             package = "tidygapminder")
xlsx_filepath <- system.file(
                 "extdata/agriculture_land.xlsx",
                 package = "tidygapminder")

tidy_indice(csv_filepath)
tidy_indice(xlsx_filepath)

}
